model-training-service/requirements.txt
pandas
scikit-learn
psycopg2-binary
xgboost
joblib
mlflow